+++ docs/changelog.md
@@ -57,6 +57,10 @@ Fixed:
   with pre-2022 BAM data (!425).
 - [CookieboxCalibration][extra.applications.CookieboxCalibration] takes polarization into account
   when calculating transmission.
+- [CookieboxCalibration][extra.applications.CookieboxCalibration] avoids crashes if eTOF data is noisy
+  by using RANSAC for the fit, using the Golden method for the t0 search, and added an option for
+  using photo-electron counting for increased robustness. Peaks are also masked if not found to avoid
+  contaminating the fit (!476).
 
 Changed: